✅ secure DELETE /tokens

+++ apps/websheep-api/src/app/token/tokens.router.ts
@@ -1,5 +1,6 @@
 import { pbkdf2Sync } from 'crypto';
 import { Router } from 'express';
+import { bearerAuthMiddleware } from '../bearer-auth-middleware';
 import { farmersService } from '../farmer/farmers.service';
 import { tokensService } from './tokens.service';
 
@@ -32,8 +33,16 @@ tokensRouter.post('/tokens', async (req, res) => {
   });
 });
 
-tokensRouter.delete('/tokens/:tokenId', (req, res) => {
+tokensRouter.delete('/tokens/:tokenId', bearerAuthMiddleware, (req, res) => {
   const { tokenId } = req.params;
+
+  const userId = tokensService.getUserId({ tokenId });
+
+  /* User should be owner of the token. */
+  if (userId !== req['user'].id) {
+    res.sendStatus(403);
+    return;
+  }
 
   tokensService.delete({ tokenId });
